refactor: consolidate CLI entry points, extract programmatic API to lib.ts (#26)
Extract library utilities from main.ts into a new src/lib.ts module that
has no Commander imports and no process-level side effects. runTests()
no longer installs signal handlers; those belong exclusively in the CLI
path. cli.ts now imports createDefaultConfig from lib instead of
maintaining a divergent inline copy. index.ts re-exports agents,
orchestrator, and the programmatic API so library consumers get the full
surface from a single import. main.ts is reduced to a thin backwards-
compatible wrapper. package.json gains an exports map restricting deep
imports to the intended entry points.
